+++ lib/lightning/projects/policy.ex
@@ -8,6 +8,11 @@ defmodule Lightning.Projects.Policy do
   alias Lightning.Projects
   alias Lightning.Projects.Project
 
+  # Users with admin level access to a project can access project settings with read/write rights
+  def authorize(:edit, %User{} = user, %Project{} = project) do
+    Projects.get_project_user_role(user, project) == :admin
+  end
+
   # Project members can read a project
   def authorize(action, user, %Project{id: _id} = project)
       when action in [:read, :index] do