+++ priv/repo/migrations/20230110173451_add_description_to_projects_table.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddDescriptionToProjectsTable do
+  use Ecto.Migration
+
+  def change do
+    alter table(:projects) do
+      add :description, :string, null: true
+    end
+  end
+end